server/requirements.txt
fastapi
uvicorn
easyocr
numpy
Pillow
opencv-python-headless
python-multipart
